Update CI jobs/search-solr-tools deploymnet
Updated jenkins job naming convention to put identifiers at the front of the title such that ordering looks as we’d expect it to on CI

Change search-solr-tools jobs to check out the project on the target machine rather than check it out on CI and rsync it on provision.  This should be far less fragile than the previous setup.